requirements.txt
transformers
fastapi
torch
uvicorn
pydantic-settings
